+++ internal/usecase/account_types.go
@@ -57,7 +57,7 @@ type AccountWalletsResBody struct {
 }
 
 type WalletSummary struct {
-	ID      string  `json:"id"`
-	Type    string  `json:"type"`
-	Balance float64 `json:"balance"`
+	ID      string `json:"id"`
+	Type    string `json:"type"`
+	Balance int64  `json:"balance"`
 }
