+++ internal/usecase/wallet_types.go
@@ -1,8 +1,8 @@
 package usecase
 
 type WalletResBody struct {
-	ID           string  `json:"id"`
-	AccountPhone string  `json:"account_phone"`
-	Type         string  `json:"type"`
-	Balance      float64 `json:"balance"`
+	ID           string `json:"id"`
+	AccountPhone string `json:"account_phone"`
+	Type         string `json:"type"`
+	Balance      int64  `json:"balance"`
 }
